Remove distutils.py
