+++ src/qibo/tests/test_models_distcircuit.py
@@ -4,7 +4,7 @@ import numpy as np
 import qibo
 from qibo import gates
 from qibo.models import Circuit
-from qibo.models.distutils import DistributedQubits
+from qibo.models.distcircuit import DistributedQubits
 
 
 def check_device_queues(queues):
